requirements.txt
streamlit>=1.35.0
langchain==0.2.14
langchain-community==0.2.12

# huggingface-hub>=0.23.0

langchain-huggingface>=0.0.3

lancedb>=0.4.3 

sentence-transformers>=2.7.0
pypdf
docx2txt
python-pptx
loguru
python-dotenv